+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Use canvas renderer for better performance
-
+- Improved popup design
 ## [0.5.0] - 2025-02-23
 
 ### Added